feat(individual-okr): spliting personal and company okrs page component
