+++ src/components/Page/Team/Members/members.tsx
@@ -77,6 +77,9 @@ export const TeamMembers = ({
     if (!isUserSidebarOpen) setSelectedUserID(undefined)
   }, [isUserSidebarOpen, setSelectedUserID])
 
+  console.log({ members })
+  console.log('members')
+
   return (
     <TeamSectionWrapper
       title={intl.formatMessage(messages.title, {
